+++ tests/conftest.py
@@ -51,6 +51,7 @@ MOCKED_URL_HTTPS = 'https+mock://requests-cache.com/text'
 MOCKED_URL_JSON = 'http+mock://requests-cache.com/json'
 MOCKED_URL_REDIRECT = 'http+mock://requests-cache.com/redirect'
 MOCKED_URL_REDIRECT_TARGET = 'http+mock://requests-cache.com/redirect_target'
+MOCKED_URL_404 = 'http+mock://requests-cache.com/nonexistent'
 MOCK_PROTOCOLS = ['mock://', 'http+mock://', 'https+mock://']
 
 AWS_OPTIONS = {
@@ -190,6 +191,11 @@ def get_mock_adapter() -> Adapter:
         headers={'Content-Type': 'text/plain'},
         text='mock redirected response',
         status_code=200,
+    )
+    adapter.register_uri(
+        ANY_METHOD,
+        MOCKED_URL_404,
+        status_code=404,
     )
     return adapter